move logging and app specific stuff out of cls.js
