+++ lib/initializers/errors.js
@@ -4,7 +4,7 @@ var _           = require("lodash");
 var domain      = require('domain');
 
 
-Initializer.add('startup', 'stex.errors', ['stex.initialize-app'], function(stex) {
+Initializer.add('startup', 'stex.errors', ['stex.initialize-app', 'stex.cls'], function(stex) {
 
   /// catch 404 and forwarding to error handler
   stex.app.use(function(req, res, next) {
