+++ lib/initializers/express.js
@@ -1,4 +1,5 @@
 var Initializer = require("../initializer");
+var uuid        = require("uuid");
 
 Initializer.add('startup', 'stex.express', ['stex.config', 'stex.logging', 'stex.db', 'stex.newrelic'], function(stex) {
   var express    = require("express");
@@ -58,5 +59,26 @@ Initializer.add('startup', 'stex.express', ['stex.config', 'stex.logging', 'stex
     });
 
     parser(req, res, next);
+  });
+
+
+  // create a transaction id for each request
+  app.use(function(req, res, next) {
+    var requestId = uuid.v4();
+    var namespace = stex.cls.getNamespace();
+
+    // wrap the events from request and response
+    namespace.bindEmitter(req);
+    namespace.bindEmitter(res);
+
+    // run following middleware in the scope of
+    // the namespace we created
+    namespace.run(function() {
+
+      // set requestId on the namespace, makes it
+      // available for all continuations
+      namespace.set('requestId', requestId);
+      next();
+    });
   });
 });
