+++ lib/initializers/logging.js
@@ -9,13 +9,16 @@ var util            = require("util");
  */
 
 
-Initializer.add('startup', 'stex.logging', ['stex.config'], function(stex) {
+Initializer.add('startup', 'stex.logging', ['stex.config', 'stex.cls'], function(stex) {
   stex.log = require("bristol");
+
   // add the request id to every metadata element logged
   stex.log.addTransform(function (element) {
     // check this is a meta data element
     if (element.file) {
       element.requestId = stex.cls.get('requestId');
+      console.log("requirestId");
+      console.log(element.requestId);
       return element;
     }
     return null;
